requirements.txt
altair==6.0.0
anyio==4.12.1
asgiref==3.11.0
attrs==25.4.0
av==16.1.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
cryptography==46.0.4
ctranslate2==4.6.3
curl_cffi==0.13.0
diff-match-patch==20241021
Django==5.2.9
django-axes==8.1.0
django-ckeditor==6.7.3
django-dbbackup==5.1.2
django-import-export==4.4.0
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-modeltranslation==0.19.19
django-recaptcha==4.1.0
et_xmlfile==2.0.0
faster-whisper==1.2.1
filelock==3.20.3
Flask==3.1.2
flatbuffers==25.12.19
frozendict==2.4.7
fsspec==2026.1.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.2
humanfriendly==10.0
idna==3.11
invoke==2.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mpmath==1.3.0
multitasking==0.0.12
narwhals==2.14.0
numpy==2.4.0
onnxruntime==1.23.2
opencc-python-reimplemented==0.1.7
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
paramiko==4.0.0
peewee==3.18.3
pillow==12.0.0
platformdirs==4.5.1
plotly==6.5.0
polib==1.2.0
protobuf==6.33.2
pyarrow==22.0.0
pycparser==2.23
pydeck==0.9.1
PyNaCl==1.6.2
pypdf==6.6.2
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
soupsieve==2.8.1
sqlparse==0.5.5
streamlit==1.52.2
stripe==14.2.0
sympy==1.14.0
tablib==3.9.0
tenacity==9.1.2
tokenizers==0.22.2
toml==0.10.2
tornado==6.5.4
tqdm==4.67.1
typer-slim==0.21.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.5
whitenoise==6.11.0
yfinance==1.0
gunicorn==23.0.0
